almost done

+++ perf.sh
@@ -1,10 +1,21 @@
 #! /bin/bash
 #exit
+cd /root/mnt
+unset pid
+ps -ef | grep perf.sh | grep -v grep 
+date
+pid=`ps -ef | grep perf.sh | grep -v grep | wc -l`
+echo $pid
+
+if [ "$pid" -gt "3" ]; then
+	echo "perf.sh already running, exiting"
+	exit
+fi
 RANDOM=$$
 /usr/bin/mysqlcheck --repair raspberry
+sleep 120
 sleep $(($RANDOM%600))
 /usr/sbin/ntpdate s1a.time.edu.cn
-cd /root/mnt
 rm -f tmp.*
 rm -f 420*
 rm -f 620*
@@ -12,7 +23,6 @@ if [ ! -f hour ]
 then
 	echo $(($RANDOM%24)) > hour
 fi
-date
 #num=$(($((0x`md5sum address |cut -d' ' -f1`))%24))
 read num < hour
 if [ $num -eq `date +%H` ];
